Gas 6 (#10)
* chore(mt-gas-6): pack time related variables in uint32

* chore(mt-gas-6): cast rewardsduration to uint256 when used in reward rate calcs

* chore(mt-gas-6): rebase

+++ src/interfaces/IArcadeStakingRewards.sol
@@ -35,7 +35,7 @@ interface IArcadeStakingRewards {
 
     function getRewardForDuration() external view returns (uint256);
 
-    function lastTimeRewardApplicable() external view returns (uint256);
+    function lastTimeRewardApplicable() external view returns (uint32);
 
     function rewardPerToken() external view returns (uint256);
 
@@ -74,7 +74,7 @@ interface IArcadeStakingRewards {
 
     function withdraw(uint256 amount, uint256 depositId) external;
 
-    function setRewardsDuration(uint256 _rewardsDuration) external;
+    function setRewardsDuration(uint32 _rewardsDuration) external;
 
     function recoverERC20(address tokenAddress, uint256 tokenAmount) external;
